Partitioned existing locale handling tests
